requirements.txt
altair==5.2.0
attrs==23.1.0
backports.zoneinfo==0.2.1
black==22.6.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.15.1
charset-normalizer==3.3.2
click==8.1.3
colorama==0.4.6
common==0.1.2
contourpy==1.1.1
cycler==0.12.1
distlib==0.3.3
fonttools==4.46.0
fsspec==2021.10.1
gitdb==4.0.11
GitPython==3.1.40
idna==3.6
importlib-metadata==6.11.0
importlib-resources==6.1.1
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.4
mdurl==0.1.2
mypy-extensions==0.4.3
numpy==1.24.3
packaging==23.2
pandas==2.0.3
pandas-read-xml==0.3.1
pathspec==0.9.0
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
platformdirs==2.5.2
plotly==5.18.0
protobuf==4.25.1
pyarrow==14.0.1
pydeck==0.8.1b0
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
referencing==0.32.0
requests==2.31.0
rich==13.7.0
rpds-py==0.13.2
rsconnect-python==1.21.0
s3fs==2021.10.1
seaborn==0.13.0
semver==2.13.0
six==1.16.0
smmap==5.0.1
streamlit==1.29.0
tableauhyperapi==0.0.15305
tableauhyperio==0.8.2
tenacity==8.2.3
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.4
typing_extensions==4.3.0
tzdata==2023.3
tzlocal==5.2
urllib3==1.26.3
validators==0.22.0
watchdog==3.0.0
weasyprint==57.2
xmltodict==0.13.0
zipp==3.17.0
